+++ forge/changelog.md
@@ -1,1 +1,2 @@
-- Updated German translation
+- Improved audio quality in bad networking conditions
+- Added FEC to opus encoder/decoder