+++ forge/changelog.md
@@ -31,3 +31,4 @@
 - Added microphone test button in microphone selection list
 - Improved native library loading and validation
 - Added `bypassGroupIsolation` to static audio channels
+- Prevent accidentally assigning left mouse button as push to talk in the voice chat settings