+++ forge/changelog.md
@@ -11,7 +11,7 @@
 - Reworked player volume adjustment
 - Allow translating volume category names and descriptions
 - Allow voice host with just port
-- Switched to 256 bit `AES/GCM/NoPadding` encryption
+- Switched to 128 bit `AES/GCM/NoPadding` encryption
 - Allow multiple targets for static audio channels
 - Removed `whisper_distance_multiplier` and `crouch_distance_multiplier` config option
 - Added `whisper_distance` config option - This now also allows for higher distances than the default voice chat distance