+++ forge/changelog.md
@@ -1,1 +1,2 @@
 - Prevent native objects from not getting closed
+- Fixed mp3 decoding issues