+++ forge/changelog.md
@@ -1,2 +1,1 @@
-- Added more detailed microphone error messages when using the java microphone implementation
-- Improved incompatible version message
+- Updated to 1.21.9